+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/caching
 
+## 0.18.33
+
+### Patch Changes
+
+- Updated dependencies [e3cf16a]
+  - @cat-factory/kernel@0.282.0
+
+## 0.18.32
+
+### Patch Changes
+
+- @cat-factory/kernel@0.281.3
+
 ## 0.18.31
 
 ### Patch Changes